Add feature to extract video frames when editing q-tube metadata
phew that was hard-won.  Phil's sample code helped a ton but there
were plenty of struggles.  in the end my testing within Firefox via
dev mode seemed to be causing frame extraction to "frequently" fail
with just a blank image.  never got to the bottom of that, but it
doesn't seem to be an issue in Chromium or Qortal UI (dev mode) so
hoping the Firefox issue can just be ignored..